LPS-57247 Move base class to a new package as well
