+++ modules/apps/journal/journal-service/src/com/liferay/journal/lar/JournalArticleExportImportContentProcessor.java
@@ -15,7 +15,7 @@
 package com.liferay.journal.lar;
 
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
-import com.liferay.exportimport.lar.BaseExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.base.BaseExportImportContentProcessor;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
