+++ modules/apps/wiki/wiki-service/src/com/liferay/wiki/lar/WikiPageExportImportContentProcessor.java
@@ -15,7 +15,7 @@
 package com.liferay.wiki.lar;
 
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
-import com.liferay.exportimport.lar.BaseExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.base.BaseExportImportContentProcessor;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
